requirements.txt
appdirs==1.4.3
beautifulsoup4==4.8.2
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
DateTime==4.3
dialogflow==0.7.2
distlib==0.3.0
docx==0.2.4
et-xmlfile==1.0.1
filelock==3.0.12
Flask==1.1.1
future==0.18.2
google-api-core==1.16.0
google-auth==1.11.2
googleapis-common-protos==1.51.0
grpcio==1.27.2
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.5.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
kiwisolver==1.1.0
line-bot-sdk==1.16.0
line-notify==0.1.4
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.1.3
MouseInfo==0.1.2
numpy==1.18.1
openpyxl==3.0.3
pandas==1.0.1
pandas-datareader==0.8.1
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyGetWindow==0.0.8
PyMsgBox==1.0.7
pyparsing==2.4.6
pyperclip==1.7.0
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==2.8.0
python-docx==0.8.10
PyTweening==1.0.3
pytz==2019.1
requests==2.23.0
rsa==4.0
selenium==3.141.0
six==1.14.0
soupsieve==2.0
urllib3==1.25.8
virtualenv==20.0.7
Werkzeug==1.0.0
zipp==3.1.0
zope.interface==4.6.0
